refactor + remove checks
